tests/wpt/tests/css/css-break/parsing/box-decoration-break-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fragmentation Module Level 3: parsing box-decoration-break with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-break/#propdef-box-decoration-break">
<meta name="assert" content="box-decoration-break supports only the grammar 'slice | clone'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("box-decoration-break", "auto");
test_invalid_value("box-decoration-break", "slice clone");
</script>
</body>
</html>
